auth improvements
